feat: added zoomed map option while editing
